Code Refactoring

+++ Entities/Concrete/City.cs
@@ -7,8 +7,8 @@ namespace Entities.Concrete
 {
     public class City:IEntity
     {
-        public long Id { get; set; }
-        public long CountryId { get; set; }
+        public int Id { get; set; }
+        public int CountryId { get; set; }
         public string Name { get; set; }
         public DateTime CreateDate { get; set; }
         public bool Active { get; set; }
