+++ Entities/Concrete/Product.cs
@@ -7,9 +7,9 @@ namespace Entities.Concrete
 {
     public class Product:IEntity
     {
-        public long Id { get; set; }
-        public long CategoryId { get; set; }
-        public long BrandId { get; set; }
+        public int Id { get; set; }
+        public int CategoryId { get; set; }
+        public int BrandId { get; set; }
         public string Name { get; set; }
         public string Code { get; set; }
         public decimal Price { get; set; }
